change mock-offers.data, refactor project structure
